% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_tiles.R
\name{gdal_tiles}
\alias{gdal_tiles}
\title{Create tiles, like gdal2tiles.py}
\usage{
gdal_tiles(
  dsn,
  zoom = NULL,
  blocksize = 256L,
  profile = "mercator",
  output_dir = tempfile(),
  overwrite = FALSE,
  update = FALSE,
  dry_run = TRUE,
  xyz = FALSE,
  format = c("png", "jpeg")
)
}
\arguments{
\item{dsn}{input dataset, file path, VRT string, or any DSN GDAL can open and warp from}

\item{zoom}{zooms to render, can be a single number multiple (from 0:23)}

\item{blocksize}{size of tiles, defaults to 256}

\item{profile}{domain to use, 'mercator', 'geodetic' (longlat), or 'raster'}

\item{output_dir}{directory to write to, by default a tempdir is used}

\item{overwrite}{clobber the output directory, \code{FALSE} is the default}

\item{update}{not implemented (please take care)}

\item{dry_run}{if \code{TRUE} only the scheme is built and returned as a data frame}

\item{xyz}{is the zero-row tile to be at the top, then set this \code{TRUE}}

\item{format}{'png' or 'jpeg'}
}
\value{
the tile scheme, invisibly as a dataframe
}
\description{
Create tiles, like gdal2tiles.py
}
\examples{
dsn <- system.file("extdata/gebco_ovr5.vrt", package = "filearchy", mustWork = TRUE)
## parallelize here
#future::plan(multicore)
tiles <- gdal_tiles(dsn, dry_run = TRUE)
if (!interactive()) unlink(tiles$path)
#future::plan(sequential)
}
