% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_tiles.R
\name{gdal_tiles}
\alias{gdal_tiles}
\title{Create tiles, like gdal2tiles.py}
\usage{
gdal_tiles(
  dsn,
  nzoom = NULL,
  blocksize = 256L,
  t_srs = "EPSG:3857",
  update = TRUE,
  output_dir = tempfile(),
  overwrite = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{dsn}{input dataset, file path, VRT string, or any DSN GDAL can open and warp from}

\item{nzoom}{number of levels, currently just single number}

\item{blocksize}{size of tiles, defaults to 256}

\item{t_srs}{crs of output, defaults to global mercator (use "EPSG:4326" for geodetic/longlat)}

\item{update}{not implemented (please take care)}

\item{output_dir}{directory to write to, by default a tempdir is used}

\item{overwrite}{clobber the output directory, \code{FALSE} is the default}

\item{dry_run}{if \code{TRUE} only the scheme is built and returned as a data frame}
}
\value{
the tile scheme, invisibly as a dataframe
}
\description{
Create tiles, like gdal2tiles.py
}
\examples{
dsn <- system.file("extdata/gebco_ovr5.vrt", package = "filearchy", mustWork = TRUE)
## parallelize here
#future::plan(multicore)
tiles <- gdal_tiles(dsn)
if (!interactive()) unlink(tiles$path)
#future::plan(sequential)
}
